Database setting to reuse query plans
